requirements.txt
flask==3.0.0
httpx==0.27.0
python-dotenv==1.0.0